+++ IntelliMower_MCU/src/utils/Angle.h
@@ -2,6 +2,6 @@
 #define ANGLE_H
 #include "Geometry.h"
 
-float GetTurnAngle(Point prvPoint, Point curPoint, Point nxtPoint);
+double GetTurnAngle(Point prvPoint, Point curPoint, Point nxtPoint);
 
 #endif //ANGLE_H
